feat() new Auth UI - Change Login UI - Change Register UI - Change primary color feat #feature/buy_book
